refactor: move SurbParams/surb_config_for/SurbConfigError to connection::options
