+++ gnosis_vpn-lib/src/route_health.rs
@@ -35,7 +35,7 @@ use std::time::{Duration, Instant, SystemTime};
 
 use crate::connection::destination::{Address, Destination, NodeId, RoutingOptions};
 use crate::connection::options::Options;
-use crate::connection::up::runner::surb_config_for;
+use crate::connection::options::surb_config_for;
 use crate::core::runner::Results;
 use crate::hopr::types::SessionClientMetadata;
 use crate::hopr::{Hopr, HoprError};
